Allow and handle zero amount invoices & LND updates

+++ src/BTCPayServer.Lightning.Eclair/Models/PayInvoiceRequest.cs
@@ -3,7 +3,7 @@ namespace BTCPayServer.Lightning.Eclair.Models
     public class PayInvoiceRequest
     {
         public string Invoice { get; set; }
-        public int? AmountMsat { get; set; }
+        public long? AmountMsat { get; set; }
         public int? MaxAttempts { get; set; }
         public int? MaxFeePct { get; set; }
         public long? MaxFeeFlatSat { get; set; }
